Use CSS instead of JS to format comment in editor toolbar (#38)
In some cases the editor toolbar does not appears, it remains blank.
This is because the loading panel is never hidden.
It happens when the loading is too long, more than the timeout set in the initBar method before calling the adjustWidth method (1500ms).

This fix replaces the way the editor toolbar display is managed. It uses CSS instead of javascript, which avoids any loading issue.
It also removes the blink effect when showing the toolbar and the editor to avoid a weird effect.